+++ src/OrganisationRegistry.Api/Backoffice/Organisation/Requests/UpdateOrganisationInfoNotLimitedToVlimpersRequest.cs
@@ -6,7 +6,6 @@
     using FluentValidation;
     using OrganisationRegistry.Organisation;
     using OrganisationRegistry.Organisation.Commands;
-    using OrganisationRegistry.Organisation.UpdateNotLimitedToVlimpers;
     using Purpose;
 
     public class UpdateOrganisationInfoNotLimitedToVlimpersInternalRequest
