+++ src/OrganisationRegistry/Organisation/Update/Events/OrganisationArticleUpdated.cs
@@ -1,0 +1,18 @@
+﻿namespace OrganisationRegistry.Organisation.Events;
+
+using System;
+
+public class OrganisationArticleUpdated : BaseEvent<OrganisationArticleUpdated>
+{
+    public Guid OrganisationId => Id;
+
+    public string? Article { get; }
+
+    public OrganisationArticleUpdated(
+        Guid organisationId,
+        string? article){
+        Id = organisationId;
+
+        Article = article;
+    }
+}
