+++ src/OrganisationRegistry/Organisation/Update/Events/OrganisationOperationalValidityUpdated.cs
@@ -1,0 +1,22 @@
+﻿namespace OrganisationRegistry.Organisation.Events;
+
+using System;
+
+public class OrganisationOperationalValidityUpdated : BaseEvent<OrganisationOperationalValidityUpdated>
+{
+    public Guid OrganisationId => Id;
+    public DateTime? OperationalValidFrom { get; }
+    public DateTime? OperationalValidTo { get; }
+
+
+    public OrganisationOperationalValidityUpdated(
+        Guid organisationId,
+        DateTime? operationalValidFrom,
+        DateTime? operationalValidTo)
+    {
+        Id = organisationId;
+
+        OperationalValidFrom = operationalValidFrom;
+        OperationalValidTo = operationalValidTo;
+    }
+}
